move all modules (each its own function) into displayPage subfolder in Screen
